Applied handlers for some error styling, character count, preventing default events and allowing default events with valid conditions
Applied handlers for some error styling: If characters in tweet text
area is greater than 140, text area color is red. If the number of
characters is 140 or below, the text color is grey.

Character count: managed with a keyup event handler.

Preventing default events: done with a check to tweet character length
and if tweet length is greater than 140, then what will be the post
request for the tweet cannot be sent.

Resets are only applied when tweet is validated, else maintain error
styles.